Added env vars, rebased creation folders func & avatar editor func. Added uniq name by user for filename
