+++ .env.example
@@ -2,3 +2,5 @@ PORT={Connection port}
 DB_HOST={MongoDB URI string}
 JWT_SECRET_KEY={secret word}
 UPLOAD_DIR={Path to the temp folder}
+PUBLIC_DIR={Path to the public folder}
+USERS_AVATARS={Path to the avatars folder}
