+++ bin/server.js
@@ -1,24 +1,11 @@
 const app = require('../app')
-const db = require('../db')
-
 const path = require('path')
-const fs = require('fs').promises
+const db = require('../db')
+const { createFoldereIsNotExist } = require('../helpers/foldersCreator')
 
 const PORT = process.env.PORT || 3000
 const UPLOAD_DIR = path.join(process.cwd(), process.env.UPLOAD_DIR)
 const AVATARS_DIR = path.join(process.cwd(), process.env.PUBLIC_DIR, process.env.USERS_AVATARS)
-
-// Проверяет права на запись (нужно вынести отдельно !!!)
-const isAccessible = async (path) => {
-  return await fs.access(path).then(() => true).catch(() => false)
-}
-
-// Создает папку, если есть права на запись (нужно вынести отдельно !!!)
-const createFoldereIsNotExist = async (folder) => {
-  if (!(await isAccessible(folder))) {
-    await fs.mkdir(folder)
-  }
-}
 
 // Старт сервера с подключением к базе
 const start = async () => {
