+++ helpers/foldersCreator.js
@@ -1,21 +1,17 @@
-// const path = require('path')
-// const fs = require('fs').promises
+const fs = require('fs').promises
 
-// const UPLOAD_DIR = path.join(process.cwd(), process.env.UPLOAD_DIR)
-// const AVATARS_DIR = path.join(process.cwd(), 'public', 'avatars')
+// Проверяет права на запись (нужно вынести отдельно в конфиг!!!)
+const isAccessible = async (path) => {
+  return await fs.access(path).then(() => true).catch(() => false)
+}
 
-// // Проверяет права на запись (нужно вынести отдельно в конфиг!!!)
-// const isAccessible = async (path) => {
-//   return await fs.access(path).then(() => true).catch(() => false)
-// }
+// Создает папку, если есть права на запись (нужно вынести отдельно в конфиг!!!)
+const createFoldereIsNotExist = async (folder) => {
+  if (!(await isAccessible(folder))) {
+    await fs.mkdir(folder)
+  }
+}
 
-// // Создает папку, если есть права на запись (нужно вынести отдельно в конфиг!!!)
-// const createFoldereIsNotExist = async (folder) => {
-//   if (!(await isAccessible(folder))) {
-//     await fs.mkdir(folder)
-//   }
-// }
-
-// module.exports = {
-//   createFoldereIsNotExist
-// }
+module.exports = {
+  createFoldereIsNotExist
+}
