+++ helpers/upload.js
@@ -1,17 +1,18 @@
 // Обрабатывает загружаемые изображения и переносит их во временную папку
 const path = require('path')
 const multer = require('multer')
-const { customAlphabet } = require('nanoid')
-const nanoid = customAlphabet('1234567890abcdef', 10)
+// const { customAlphabet } = require('nanoid')
+// const nanoid = customAlphabet('1234567890abcdef', 10)
 
 const UPLOAD_DIR = path.join(process.cwd(), process.env.UPLOAD_DIR)
 
 const storage = multer.diskStorage({
   destination: function (req, file, cb) {
+    console.log(req.user.id)
     cb(null, UPLOAD_DIR)
   },
   filename: function (req, file, cb) {
-    cb(null, nanoid() + path.extname(file.originalname))
+    cb(null, `${req.user.id}${path.extname(file.originalname)}`)
   }
 })
