+++ services/userService.js
@@ -29,7 +29,7 @@ const updateSubscription = async (id, subscription) => {
   return user
 }
 
-// Обновляет аватар юзера - удалить старый аватар!!!
+// Обновляет аватар юзера
 const updateAvatar = async (id, url) => {
   const { avatarURL } = await User.findOneAndUpdate({ _id: id }, { avatarURL: url }, { new: true }) // Обновляет аватар в базе
